Extract the tag list to it's own file.
